Termination Console now doesn't work if the admin logs out and leaves it up
Started work on making the job images changable

+++ Los Alamos Timeclock/UI/TerminationConsole.cs
@@ -27,22 +27,31 @@ namespace Los_Alamos_Timeclock.UI
 
         private void terminateButton_Click(object sender, EventArgs e)
         {
-            Main.maininstance.sqlCommand("UPDATE Employee SET EDate='" + lastDayCalander.Value.ToString("yyyy-MM-dd") + "', EReason='" + reasonDropdownlist.Text + "' WHERE ID='" + id + "'");
-            String logString = reasonDropdownlist.Text.ToUpper() + ": " + commentsTextbox.Text.Replace(@"\", @"\\").Replace("'", @"\'") + "\nLast Day: " + lastDayCalander.Value.ToShortDateString();
-            Main.maininstance.sqlCommand("INSERT INTO EmployeeNotes VALUES('" + id + "','" + Main.eName.Replace(@"\", @"\\").Replace("'", @"\'") + "', NOW() ,'" + logString.Replace(@"\", @"\\").Replace("'", @"\'") + "')");
-            Main.maininstance.sqlCommand("DELETE FROM Schedule WHERE ID='" + id + "' and Date>'" + lastDayCalander.Value.ToString("yyyy-MM-dd") + "'");
-
-            if (removePrivCheckbox.Checked)
+            if (Main.permissions == "Admin")
             {
-                Main.maininstance.sqlCommand("DELETE FROM Admin WHERE ID='" + id + "'");
-                Main.maininstance.sqlCommand("DELETE FROM Manager WHERE ID='" + id + "'");
+
+                Main.maininstance.sqlCommand("UPDATE Employee SET EDate='" + lastDayCalander.Value.ToString("yyyy-MM-dd") + "', EReason='" + reasonDropdownlist.Text + "' WHERE ID='" + id + "'");
+                String logString = reasonDropdownlist.Text.ToUpper() + ": " + commentsTextbox.Text.Replace(@"\", @"\\").Replace("'", @"\'") + "\nLast Day: " + lastDayCalander.Value.ToShortDateString();
+                Main.maininstance.sqlCommand("INSERT INTO EmployeeNotes VALUES('" + id + "','" + Main.eName.Replace(@"\", @"\\").Replace("'", @"\'") + "', NOW() ,'" + logString.Replace(@"\", @"\\").Replace("'", @"\'") + "')");
+                Main.maininstance.sqlCommand("DELETE FROM Schedule WHERE ID='" + id + "' and Date>'" + lastDayCalander.Value.ToString("yyyy-MM-dd") + "'");
+
+                if (removePrivCheckbox.Checked)
+                {
+                    Main.maininstance.sqlCommand("DELETE FROM Admin WHERE ID='" + id + "'");
+                    Main.maininstance.sqlCommand("DELETE FROM Manager WHERE ID='" + id + "'");
+                }
+                Main.maininstance.getEmployees();
+                MessageBox.Show("Employee Terminated");
+                Main.maininstance.panel1.Controls.Clear();
+                Main.maininstance.panel1.Controls.Add(new Editemployees());
+                Main.maininstance.panel1.Controls[0].Dock = DockStyle.Fill;
+                this.Close();
             }
-            Main.maininstance.getEmployees();
-            MessageBox.Show("Employee Terminated");
-            Main.maininstance.panel1.Controls.Clear();
-            Main.maininstance.panel1.Controls.Add(new Editemployees());
-            Main.maininstance.panel1.Controls[0].Dock = DockStyle.Fill;
-            this.Close();
+            else
+            {
+                MessageBox.Show("Not an admin");
+                this.Close();
+            }
         }
     }
 }
